requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.20.0
MLB-StatsAPI>=1.6.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pybaseball>=2.2.0
python-dotenv>=1.0.0
pytz>=2024.1
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.0.0
